+++ plugins/test/myCustomHeader.vue
@@ -1,6 +1,9 @@
 <template>
   <div>
-    Ha, this is a custom Header with a <span @click="clicked"> CLOSE thing </span>
+    <div>
+      Ha, this is a custom Header with a <span @click="clicked"> CLOSE thing. </span>
+    </div>
+    <div>And here is '<b>{{ sample }}</b>' passed through</div>
   </div>
 </template>
 
@@ -8,6 +11,11 @@
 
   export default {
     name: 'MyCustomHeader',
+    props: {
+      sample: {
+        type: String,
+      },
+    },
     setup(props, { emit }) {
       return {
         clicked: () => {