+++ src/manager/window/WindowComponentHeader.vue
@@ -10,30 +10,55 @@
         {{ $t(windowState.headerTitle) }}
       </h3>
     </span>
-    <VcsButton
-      @click.stop="close"
-      small
-      icon="mdi-close-thick"
-      tooltip="components.close"
-    />
+    <div class="d-flex justify-space-between align-center">
+      <template v-if="windowState.headerActions?.length > 0">
+        <VcsActionButtonList
+          :actions="windowState.headerActions"
+          :overflow-count="windowState.headerActionsOverflowCount ?? 3"
+          small
+        />
+        <v-divider
+          vertical
+          inset
+          class="mx-2"
+        />
+      </template>
+      <VcsButton
+        @click.stop="close"
+        small
+        icon="mdi-close-thick"
+        tooltip="components.close"
+      />
+    </div>
   </span>
 </template>
 
-<style>
+<style lang="scss" scoped>
 .window-component-header{
   max-height: 16px;
+
+  .v-divider--vertical.v-divider--inset {
+    margin-top: 2px;
+  }
 }
 </style>
 
 <script>
-  import { VIcon } from 'vuetify/lib';
+  import { VIcon, VDivider } from 'vuetify/lib';
   import VcsButton from '../../components/buttons/VcsButton.vue';
+  import VcsActionButtonList from '../../components/buttons/VcsActionButtonList.vue';
 
+  /**
+   * Default window component header with drag functionality close action and further optional window actions.
+   * @vue-prop {WindowState}    windowState - state of the window component.
+   */
   export default {
     name: 'WindowComponentHeader',
     components: {
+      VcsActionButtonList,
       VcsButton,
       VIcon,
+      VDivider,
     },
     props: {
       windowState: {