+++ src/manager/window/WindowManager.vue
@@ -23,6 +23,7 @@
         <component
           :is="getHeaderComponent(id)"
           :window-state="getState(id)"
+          v-bind="getProps(id)"
           @close="close(id)"
         />
       </template>